Moxe metadata properties to more suitable places
