+++ src/layman/layer/micka/util_test.py
@@ -15,6 +15,8 @@ from layman.layer import LAYER_TYPE
 from .csw import _get_template_values
 
 from layman.common.micka import util as common_util
+from layman.common.metadata import PROPERTIES as COMMON_PROPERTIES, prop_equals
+from .csw import METADATA_PROPERTIES
 
 
 @pytest.fixture(scope="module")
@@ -59,25 +61,6 @@ def test_fill_template(client):
     except OSError:
         pass
     file_object = common_util.fill_template('src/layman/layer/micka/record-template.xml', _get_template_values())
-    with open(xml_path, 'w') as out:
-        out.write(file_object.read())
-
-    def get_diff(p1, p2):
-        diff = difflib.unified_diff(open(p1).readlines(), open(p2).readlines())
-        return f"diff={''.join(diff)}"
-
-    expected_path = 'src/layman/layer/micka/util_test_filled_template.xml'
-    assert filecmp.cmp(xml_path, expected_path, shallow=False), get_diff(xml_path, expected_path)
-
-
-@pytest.mark.usefixtures('app_context')
-def test_fill_template(client):
-    xml_path = 'tmp/record-template.xml'
-    try:
-        os.remove(xml_path)
-    except OSError:
-        pass
-    file_object = common_util.fill_template('src/layman/layer/micka/record-template.xml', _get_template_values())
     with open(xml_path, 'wb') as out:
         out.write(file_object.read())
 
@@ -87,6 +70,46 @@ def test_fill_template(client):
 
     expected_path = 'src/layman/layer/micka/util_test_filled_template.xml'
     assert filecmp.cmp(xml_path, expected_path, shallow=False), get_diff(xml_path, expected_path)
+
+
+def test_parse_md_properties():
+    xml_path = 'src/layman/layer/micka/util_test_filled_template.xml'
+    with open(xml_path, 'r') as xml_file:
+        props = common_util.parse_md_properties(xml_file, [
+           'abstract',
+           'extent',
+           'graphic_url',
+           'identifier',
+           'layer_endpoint',
+           'md_date_stamp',
+           'md_file_identifier',
+           'publication_date',
+           'reference_system',
+           'title',
+           'wfs_url',
+           'wms_url',
+        ], METADATA_PROPERTIES)
+    expected = {
+        'md_file_identifier': 'm-ca238200-8200-1a23-9399-42c9fca53542',
+        'md_date_stamp': '2007-05-25',
+        'reference_system': [4326, 3857],
+        'title': 'CORINE - Krajinný pokryv CLC 90',
+        'publication_date': '2007-05-25',
+        'identifier': {
+            'identifier': 'http://www.env.cz/data/corine/1990',
+            'label': 'MZP-CORINE',
+        },
+        'abstract': None,
+        'graphic_url': 'http://layman_test_run_1:8000/rest/browser/layers/layer/thumbnail',
+        'extent': [11.87, 48.12, 19.13, 51.59],
+        'wms_url': 'http://www.env.cz/corine/data/download.zip',
+        'wfs_url': 'http://www.env.cz/corine/data/download.zip',
+        'layer_endpoint': 'http://layman_test_run_1:8000/rest/browser/layers/layer',
+    }
+    assert set(props.keys()) == set(expected.keys())
+    for k in props.keys():
+        equals_fn = COMMON_PROPERTIES[k].get('equals_fn', None)
+        assert prop_equals(props[k], expected[k], equals_fn), f"Values of property {k} do not equal: {props[k]} != {expected[k]}"
 
 
 @pytest.mark.usefixtures('app_context')
@@ -112,7 +135,7 @@ def test_fill_xml_template(client):
         'wms_url': 'https://example.com/wms',
         'wfs_url': 'https://example.com/wfs',
         'layer_endpoint': 'https://example.com/layer_endpoint',
-    })
+    }, METADATA_PROPERTIES)
 
     expected_path = 'src/layman/layer/micka/simple-test-template-filled.xml'
     with open(expected_path) as f:
